Use routers to clean up code; organize textGear API calls; Save API calls in txt files
